pyproject.toml
[project]
name = "groq-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.17.0",
    "groq>=0.31.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-cohere>=0.4.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
